requirements.txt
asyncpraw==7.8.1
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
pyyaml==6.0.1
schedule==1.2.0
praw
asyncpg==0.29.0 